+++ src/Html/Column.php
@@ -5,6 +5,7 @@ namespace Yajra\DataTables\Html;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 use Illuminate\Support\Fluent;
+use Yajra\DataTables\Html\Options\Plugins\SearchPanes;
 
 /**
  * @property string data
@@ -19,6 +20,8 @@ use Illuminate\Support\Fluent;
  */
 class Column extends Fluent
 {
+    use SearchPanes;
+
     /**
      * @param array $attributes
      */
@@ -101,7 +104,7 @@ class Column extends Fluent
 
         return $this;
     }
-    
+
      /**
      * Set column responsive priority.
      *
@@ -142,6 +145,23 @@ class Column extends Fluent
         $attr = [
             'data' => $data,
             'name' => $name ?: $data,
+        ];
+
+        return new static($attr);
+    }
+
+    /**
+     * Make a new formatted column instance.
+     *
+     * @param string $name
+     * @return Column
+     */
+    public static function formatted($name)
+    {
+        $attr = [
+            'data'  => $name . '_formatted',
+            'name'  => $name,
+            'title' => self::titleFormat($name),
         ];
 
         return new static($attr);