+++ src/Html/ColumnDefinitions.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Yajra\DataTables\Html;
+
+use Illuminate\Support\Collection;
+
+class ColumnDefinitions extends Collection
+{
+
+}