+++ CHANGELOG.md
@@ -8,6 +8,14 @@
 
 ## CHANGELOG
 
+### v4.35.0 - 11-03-2020
+
+- Add formatted column factory. [#147]
+
+### v4.34.0 - 10-31-2020
+
+- Add support for search panes extension. [#137]
+
 ### v4.33.0 - 10-30-2020
 
 - Make LaravelDataTables javascript namespace configurable. [#145], credits to @om3rcitak
@@ -699,6 +707,8 @@ To `created_at` with title `Created At`
 [#142]: https://github.com/yajra/laravel-datatables-html/pull/142
 [#143]: https://github.com/yajra/laravel-datatables-html/pull/143
 [#144]: https://github.com/yajra/laravel-datatables-html/pull/144
+[#137]: https://github.com/yajra/laravel-datatables-html/pull/137
+[#147]: https://github.com/yajra/laravel-datatables-html/pull/147
 
 [#134]: https://github.com/yajra/laravel-datatables-html/issues/134
 [#3]: https://github.com/yajra/laravel-datatables-html/issues/3